Project1
Given a composite system comprising of randomly distributed insulating and metallic materials: what fraction of the system
needs to be metallic so that the composite system is an electrical conductor? Given a porous landscape with water on the
surface (or oil below), under what conditions will the water be able to drain through to the bottom (or the oil to gush through
to the surface)? Scientists have dened an abstract process known as percolation to model such situations.